refactor(cli): extract merge_config and its tests into config_resolver.rs
Move the merge_config function (~161 lines) and all merge_config-related
tests (~200 lines) from main.rs into src/cli/config_resolver.rs.

- Add UnknownLicenseHandling to config_resolver.rs imports
- Make merge_config pub in config_resolver.rs
- Add clap::Parser import in test module for Args::parse_from
- Update main.rs to import merge_config from config_resolver
- Remove now-unused imports (LicensePolicy, UnknownLicenseHandling,
  Severity, ConfigFile, IgnoreCve, MergedConfig) from main.rs
- main.rs now contains only module declarations, run(), and main()

main.rs is now ≤200 lines (non-blank/non-comment), completing #315.

Closes #355
